refactor: Move debug utility
