--- src/debug.ts
@@ -1,3 +1,0 @@
-import createDebug from 'debug'
-
-export default createDebug('vue-codemod')
