+++ src/transformTemplate.ts
@@ -1,7 +1,7 @@
-import debug from '~/debug'
 import processTransformResult from '~/processTransformResult'
 import type { TemplateTransformation } from '~/types/TemplateTransformation'
 import type { TransformationBlock } from '~/types/TransformationBlock'
+import debug from '~/utils/debug'
 
 export default function transformTemplate(
   transformation: TemplateTransformation,
