+++ src/utils/debug.ts
@@ -1,0 +1,5 @@
+import createDebug from 'debug'
+
+export default createDebug('vue-sfcmod')
+
+export const debugTemplate = createDebug('vue-sfcmod:template')
